Refactor: Relocate Classes and Update Packages
- Move `PathData` from `mvi` package to `model` package.
- Move `Color.kt`, `Theme.kt`, and `Type.kt` files to `core/presentation/ui/theme` package.
- Move `MainActivity` to `core/presentation` package.
- Move `DifficultyLevel` to the `domain` package.
- Update all usages of relocated classes with new packages.
