+++ app/src/main/java/com/github/ksalil/scribbledash/core/navigation/destinations/DrawScreenDestination.kt
@@ -1,6 +1,6 @@
 package com.github.ksalil.scribbledash.core.navigation.destinations
 
-import com.github.ksalil.scribbledash.game.DifficultyLevel
+import com.github.ksalil.scribbledash.game.domain.DifficultyLevel
 import kotlinx.serialization.Serializable
 
 @Serializable
