+++ app/src/main/java/com/github/ksalil/scribbledash/core/presentation/MainActivity.kt
@@ -1,18 +1,11 @@
-package com.github.ksalil.scribbledash
+package com.github.ksalil.scribbledash.core.presentation
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.tooling.preview.Preview
 import com.github.ksalil.scribbledash.core.navigation.AppNavigation
-import com.github.ksalil.scribbledash.ui.theme.ScribbleDashTheme
+import com.github.ksalil.scribbledash.core.presentation.ui.theme.ScribbleDashTheme
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
