+++ app/src/main/java/com/github/ksalil/scribbledash/core/presentation/components/TitleWithDescription.kt
@@ -11,8 +11,8 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
-import com.github.ksalil.scribbledash.ui.theme.OnBackgroundVariant
-import com.github.ksalil.scribbledash.ui.theme.ScribbleDashTheme
+import com.github.ksalil.scribbledash.core.presentation.ui.theme.OnBackgroundVariant
+import com.github.ksalil.scribbledash.core.presentation.ui.theme.ScribbleDashTheme
 
 @Composable
 fun TitleWithDescription(
