+++ app/src/main/java/com/github/ksalil/scribbledash/core/presentation/ui/theme/Color.kt
@@ -1,4 +1,4 @@
-package com.github.ksalil.scribbledash.ui.theme
+package com.github.ksalil.scribbledash.core.presentation.ui.theme
 
 import androidx.compose.ui.graphics.Color
