+++ app/src/main/java/com/github/ksalil/scribbledash/core/presentation/ui/theme/Type.kt
@@ -1,4 +1,4 @@
-package com.github.ksalil.scribbledash.ui.theme
+package com.github.ksalil.scribbledash.core.presentation.ui.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
