+++ app/src/main/java/com/github/ksalil/scribbledash/home/presentation/HomeScreen.kt
@@ -31,9 +31,9 @@ import androidx.compose.ui.unit.dp
 import com.github.ksalil.scribbledash.R
 import com.github.ksalil.scribbledash.core.presentation.components.GameModeItem
 import com.github.ksalil.scribbledash.core.presentation.components.TitleWithDescription
-import com.github.ksalil.scribbledash.ui.theme.BackgroundGradientEnd
-import com.github.ksalil.scribbledash.ui.theme.BottomNavBarItemUnselectedColor
-import com.github.ksalil.scribbledash.ui.theme.ScribbleDashTheme
+import com.github.ksalil.scribbledash.core.presentation.ui.theme.BackgroundGradientEnd
+import com.github.ksalil.scribbledash.core.presentation.ui.theme.BottomNavBarItemUnselectedColor
+import com.github.ksalil.scribbledash.core.presentation.ui.theme.ScribbleDashTheme
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
